requirements.txt
chex
blackjax-nightly==0.9.6.post153
distrax
dm-haiku
optax
pytest
tqdm
matplotlib
pandas
jupyterlab
wandb
hydra-core
hydra-joblib-launcher